requirements.txt
tensorboard
tensorboardX
torchmetrics
opencv-python
scikit-image
pillow
numpy
matplotlib
tqdm
einops
safetensors
huggingface-hub
transformers
